new changes!!
added a radius encoder, plus did a few changes so if we want, we can add more features in the future, also added 2 new metrics for radius (avg difference and std)"
[add-radius-feat 974f2af] added a radius encoder, plus did a few changes so if we want, we can add more features in the future, also added 2 new metrics for radius (avg difference and std)

+++ vascular_network/dataset.py
@@ -239,7 +239,7 @@ class VesselGraphDataset(InMemoryDataset):
 
             # includes all edges (train+test+val) in both drections
             data.edge_index_undirected = torch.tensor(edge_array, dtype=torch.long)
-            data.edge_attr_undirected = torch.tensor(edge_attr_array, dtype=torch.long)
+            data.edge_attr_undirected = torch.tensor(edge_attr_array, dtype=torch.float)
             del edge_array
             del edge_attr_array
